+++ src/main/java/techcourse/myblog/web/controller/ArticleController.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.web;
+package techcourse.myblog.web.controller;
 
 
 import lombok.RequiredArgsConstructor;
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.*;
 import techcourse.myblog.dto.ArticleDto;
 import techcourse.myblog.service.ArticleService;
 import techcourse.myblog.domain.Comment;
+import techcourse.myblog.web.UserSession;
 
 import java.util.List;
