+++ src/test/java/techcourse/myblog/web/controller/BaseControllerTests.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.web;
+package techcourse.myblog.web.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -6,7 +6,6 @@ import org.springframework.test.web.reactive.server.WebTestClient;
 
 import static org.springframework.web.reactive.function.BodyInserters.fromFormData;
 
-@SpringBootTest
 public abstract class BaseControllerTests {
     protected static final String EMAIL_KEY = "email";
     protected static final String USER_NAME_KEY = "name";
